+++ src/app/admin/user/page.tsx
@@ -1,12 +1,15 @@
 "use client";
-import CreateModal from './components/CreateModal';
-import UpdateModal from './components/UpdateModal';
-import { deleteUserUsingPost, listUserByPageUsingPost } from '@/api/userController';
-import { PlusOutlined } from '@ant-design/icons';
-import type { ActionType, ProColumns } from '@ant-design/pro-components';
-import { PageContainer, ProTable } from '@ant-design/pro-components';
-import { Button, message, Space, Typography } from 'antd';
-import React, { useRef, useState } from 'react';
+import CreateModal from "./components/CreateModal";
+import UpdateModal from "./components/UpdateModal";
+import {
+  deleteUserUsingPost,
+  listUserByPageUsingPost,
+} from "@/api/userController";
+import { PlusOutlined } from "@ant-design/icons";
+import type { ActionType, ProColumns } from "@ant-design/pro-components";
+import { PageContainer, ProTable } from "@ant-design/pro-components";
+import { Button, message, Modal, Space, Typography } from "antd";
+import React, { useRef, useState } from "react";
 
 /**
  * 用户管理页面
@@ -28,21 +31,29 @@ const UserAdminPage: React.FC = () => {
    * @param row
    */
   const handleDelete = async (row: API.User) => {
-    const hide = message.loading('正在删除');
     if (!row) return true;
-    try {
-      await deleteUserUsingPost({
-        id: row.id as any,
-      });
-      hide();
-      message.success('删除成功');
-      actionRef?.current?.reload();
-      return true;
-    } catch (error: any) {
-      hide();
-      message.error('删除失败，' + error.message);
-      return false;
-    }
+
+    // 显示确认对话框
+    Modal.confirm({
+      title: "确认删除",
+      content: "确定要删除该用户吗？",
+      onOk: async () => {
+        try {
+          await deleteUserUsingPost({
+            id: row.id as any,
+          });
+          message.success("删除成功");
+          actionRef?.current?.reload();
+          return true;
+        } catch (error: any) {
+          message.error("删除失败，" + error.message);
+          return false;
+        }
+      },
+      onCancel() {
+        message.info("已取消删除");
+      },
+    });
   };
 
   /**
@@ -50,148 +61,155 @@ const UserAdminPage: React.FC = () => {
    */
   const columns: ProColumns<API.User>[] = [
     {
-      title: 'id',
-      dataIndex: 'id',
-      valueType: 'text',
+      title: "id",
+      dataIndex: "id",
+      valueType: "text",
       hideInForm: true,
     },
     {
-      title: '账号',
-      dataIndex: 'userAccount',
-      valueType: 'text',
+      title: "账号",
+      dataIndex: "userAccount",
+      valueType: "text",
     },
     {
-      title: '用户名',
-      dataIndex: 'userName',
-      valueType: 'text',
+      title: "用户名",
+      dataIndex: "userName",
+      valueType: "text",
     },
     {
-      title: '头像',
-      dataIndex: 'userAvatar',
-      valueType: 'image',
+      title: "头像",
+      dataIndex: "userAvatar",
+      valueType: "image",
       fieldProps: {
         width: 64,
       },
       hideInSearch: true,
     },
     {
-      title: '简介',
-      dataIndex: 'userProfile',
-      valueType: 'textarea',
+      title: "简介",
+      dataIndex: "userProfile",
+      valueType: "textarea",
     },
     {
-      title: '权限',
-      dataIndex: 'userRole',
+      title: "权限",
+      dataIndex: "userRole",
       valueEnum: {
         user: {
-          text: '用户',
+          text: "普通用户",
         },
         admin: {
-          text: '管理员',
+          text: "管理员",
+        },
+        vip: {
+          text: "VIP",
         },
       },
     },
     {
-      title: '创建时间',
+      title: "创建时间",
       sorter: true,
-      dataIndex: 'createTime',
-      valueType: 'dateTime',
+      dataIndex: "createTime",
+      valueType: "dateTime",
       hideInSearch: true,
       hideInForm: true,
     },
     {
-      title: '更新时间',
+      title: "更新时间",
       sorter: true,
-      dataIndex: 'updateTime',
-      valueType: 'dateTime',
+      dataIndex: "updateTime",
+      valueType: "dateTime",
       hideInSearch: true,
       hideInForm: true,
     },
     {
-      title: '操作',
-      dataIndex: 'option',
-      valueType: 'option',
+      title: "操作",
+      dataIndex: "option",
+      valueType: "option",
       render: (_, record) => (
-          <Space size="middle">
-            <Typography.Link
-                onClick={() => {
-                  setCurrentRow(record);
-                  setUpdateModalVisible(true);
-                }}
-            >
-              修改
-            </Typography.Link>
-            <Typography.Link type="danger" onClick={() => handleDelete(record)}>
-              删除
-            </Typography.Link>
-          </Space>
+        <Space size="middle">
+          <Button
+            style={{ color: "rgba(26,78,171,0.94)" }}
+            onClick={() => {
+              setCurrentRow(record);
+              setUpdateModalVisible(true);
+            }}
+          >
+            修改
+          </Button>
+          <Button type="dashed" danger onClick={() => handleDelete(record)}>
+            删除
+          </Button>
+        </Space>
       ),
     },
   ];
   return (
-      <PageContainer>
-        <ProTable<API.User>
-            headerTitle={'查询表格'}
-            actionRef={actionRef}
-            rowKey="key"
-            search={{
-              labelWidth: 120,
+    <PageContainer>
+      <ProTable<API.User>
+        headerTitle={"用户管理表格"}
+        actionRef={actionRef}
+        rowKey="key"
+        search={{
+          labelWidth: 120,
+        }}
+        toolBarRender={() => [
+          <Button
+            type="primary"
+            key="primary"
+            onClick={() => {
+              setCreateModalVisible(true);
             }}
-            toolBarRender={() => [
-              <Button
-                  type="primary"
-                  key="primary"
-                  onClick={() => {
-                    setCreateModalVisible(true);
-                  }}
-              >
-                <PlusOutlined /> 新建
-              </Button>,
-            ]}
-            request={async (params, sort, filter) => {
-              const sortField = Object.keys(sort)?.[0];
-              const sortOrder = sort?.[sortField] ?? undefined;
+          >
+            <PlusOutlined /> 新建
+          </Button>,
+        ]}
+        request={async (params, sort, filter) => {
+          const sortField = Object.keys(sort)?.[0];
+          const sortOrder = sort?.[sortField] ?? undefined;
 
-              const { data, code } = await listUserByPageUsingPost({
-                ...params,
-                sortField,
-                sortOrder,
-                ...filter,
-              } as API.UserQueryRequest);
+          const { data, code } = await listUserByPageUsingPost({
+            ...params,
+            sortField,
+            sortOrder,
+            ...filter,
+          } as API.UserQueryRequest);
 
-              return {
-                success: code === 0,
-                data: data?.records || [],
-                total: Number(data?.total) || 0,
-              };
-            }}
-            columns={columns}
-        />
-        <CreateModal
-            visible={createModalVisible}
-            columns={columns}
-            onSubmit={() => {
-              setCreateModalVisible(false);
-              actionRef.current?.reload();
-            }}
-            onCancel={() => {
-              setCreateModalVisible(false);
-            }}
-        />
-        <UpdateModal
-            visible={updateModalVisible}
-            columns={columns}
-            oldData={currentRow}
-            onSubmit={() => {
-              setUpdateModalVisible(false);
-              setCurrentRow(undefined);
-              actionRef.current?.reload();
-            }}
-            onCancel={() => {
-              setUpdateModalVisible(false);
-            }}
-        />
-      </PageContainer>
+          return {
+            success: code === 0,
+            data: data?.records || [],
+            total: Number(data?.total) || 0,
+          };
+        }}
+        columns={columns}
+        pagination={{
+          pageSize: 6, // 每页显示5条记录
+        }}
+      />
+      <CreateModal
+        visible={createModalVisible}
+        columns={columns}
+        onSubmit={() => {
+          setCreateModalVisible(false);
+          actionRef.current?.reload();
+        }}
+        onCancel={() => {
+          setCreateModalVisible(false);
+        }}
+      />
+      <UpdateModal
+        visible={updateModalVisible}
+        columns={columns}
+        oldData={currentRow}
+        onSubmit={() => {
+          setUpdateModalVisible(false);
+          setCurrentRow(undefined);
+          actionRef.current?.reload();
+        }}
+        onCancel={() => {
+          setUpdateModalVisible(false);
+        }}
+      />
+    </PageContainer>
   );
 };
 export default UserAdminPage;
